refactor: 폴더구조 재설정
