--- src/app/productlist/page.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const ProductList = () => {
-	return (
-		<div>
-			It is productList
-		</div>
-	);
-};
-
-export default ProductList;
